authorization continued (DAY 4 into nestjs)

+++ src/tasks/task.service.ts
@@ -5,6 +5,7 @@ import { SearchDto } from './dto/search.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
+import { User } from 'src/auth/user.entity';
 
 
 @Injectable()
@@ -15,30 +16,30 @@ export class TaskService {
         private taskRepository : TaskRepository
     ){}
 
-    async getTasks(searchDto : SearchDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(searchDto)
+    async getTasks(searchDto : SearchDto, user:User):Promise<Task[]>{
+        return this.taskRepository.getTasks(searchDto,user)
     }
 
-    async getTaskById(id:number){
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id:number,user:User){
+        const found = await this.taskRepository.findOne({where:{id , userId:user.id}});
         if(!found){
             throw new NotFoundException(`task with ID ${id} not found`);
         }
         return found
     }
 
-    async addTask(createTaskDto:CreateTaskDto):Promise<Task>{
-        return this.taskRepository.addTask(createTaskDto);
+    async addTask(createTaskDto:CreateTaskDto,user:User):Promise<Task>{
+        return this.taskRepository.addTask(createTaskDto,user);
     }
 
-    async deleteTask(id : number):Promise<void>{
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id : number,user:User):Promise<void>{
+        const result = await this.taskRepository.delete({id,userId:user.id});
     }
     
 
-    async updateTask(id:number , createTaskDto:CreateTaskDto):Promise<Task>{
+    async updateTask(id:number , createTaskDto:CreateTaskDto,user:User):Promise<Task>{
         const {title , description} = createTaskDto ;
-        const task = await this.getTaskById(id);
+        const task = await this.getTaskById(id,user);
 
         if(title) task.title = title;
         if(description) task.description = description ;
@@ -48,8 +49,8 @@ export class TaskService {
         return task;
     }
 
-    async updateTaskStatus(id:number , taskStatus:TaskStatus):Promise<Task>{
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id:number , taskStatus:TaskStatus,user:User):Promise<Task>{
+        const task = await this.getTaskById(id,user);
         task.status = taskStatus ;
         await task.save();
